[Test] Add LoRA e2e test
